feat: classes added api class updated
