+++ src/components/api.js
@@ -1,82 +1,74 @@
-const config = {
-  url: "https://nomoreparties.co/v1/plus-cohort-13",
-  headers: {
-    authorization: "48267a8c-35ef-4976-8b19-19133ce8e68c",
-    "Content-Type": "application/json",
-  },
-};
+export default class Api {
+  constructor({ url, headers }) {
+    // тело конструктора
+    this._url = url;
+    this._headers = headers;
+  }
 
-const onResponce = (responce) => {
-  return responce.ok ? responce.json() : Promise.reject(responce);
-};
+  _onResponce(responce) {
+    return responce.ok ? responce.json() : Promise.reject(responce);
+  }
 
-//загрузка всех карточек с сервера
-function getCards() {
-  return fetch(`${config.url}/cards`, {
-    headers: config.headers,
-  }).then(onResponce);
+  //методы работы с API
+
+  //загрузка всех карточек с сервера
+  getCards() {
+    return fetch(`${this._url}/cards`, {
+      headers: this._headers,
+    }).then(this._onResponce);
+  }
+
+  //получение информации о пользователе
+  getUser() {
+    return fetch(`${this._url}/users/me`, {
+      headers: this._headers,
+    }).then(this._onResponce);
+  }
+
+  getAllInfo() {
+    return Promise.all([this.getCards(), this.getUser()]);
+  }
+
+  //редактирование информации о пользователе
+  editProfile(data) {
+    return fetch(`${this._url}/users/me`, {
+      method: "PATCH",
+      headers: this._headers,
+      body: JSON.stringify(data),
+    }).then(this._onResponce);
+  }
+
+  //добавление новой карточки
+  addCard(data) {
+    return fetch(`${this._url}/cards`, {
+      method: "POST",
+      headers: this._headers,
+      body: JSON.stringify(data),
+    }).then(this._onResponce);
+  }
+
+  //удаление карточки
+  removeCard(dataId) {
+    return fetch(`${this._url}/cards/${dataId}`, {
+      method: "DELETE",
+      headers: this._headers,
+    }).then(this._onResponce);
+  }
+
+  //изменение аватара
+  editUserAvatar(data) {
+    return fetch(`${this._url}/users/me/avatar`, {
+      method: "PATCH",
+      headers: this._headers,
+      body: JSON.stringify(data),
+    }).then(this._onResponce);
+  }
+
+  // изменение состояния лайка
+  changeLikeStatus(dataId, isLiked) {
+    return fetch(`${this._url}/cards/likes/${dataId}`, {
+      method: isLiked ? "DELETE" : "PUT",
+      headers: this._headers,
+    }).then(this._onResponce);
+  }
 }
-
-//получение информации о пользователе
-function getUser() {
-  return fetch(`${config.url}/users/me`, {
-    headers: config.headers,
-  }).then(onResponce);
-}
-
-function getAllInfo() {
-  return Promise.all([getCards(), getUser()]);
-}
-
-//редактирование информации о пользователе
-function editProfile(data) {
-  return fetch(`${config.url}/users/me`, {
-    method: "PATCH",
-    headers: config.headers,
-    body: JSON.stringify(data),
-  }).then(onResponce);
-}
-
-//добавление новой карточки
-function addCard(data) {
-  return fetch(`${config.url}/cards`, {
-    method: "POST",
-    headers: config.headers,
-    body: JSON.stringify(data),
-  }).then(onResponce);
-}
-
-//удаление карточки
-function removeCard(dataId) {
-  return fetch(`${config.url}/cards/${dataId}`, {
-    method: "DELETE",
-    headers: config.headers,
-  }).then(onResponce);
-}
-
-//изменение аватара
-function editUserAvatar(data) {
-  return fetch(`${config.url}/users/me/avatar`, {
-    method: "PATCH",
-    headers: config.headers,
-    body: JSON.stringify(data),
-  }).then(onResponce);
-}
-
-// изменение состояния лайка
-function changeLikeStatus(dataId, isLiked) {
-  return fetch(`${config.url}/cards/likes/${dataId}`, {
-    method: isLiked ? "DELETE" : "PUT",
-    headers: config.headers,
-  }).then(onResponce);
-}
-
-export {
-  getCards,
-  addCard,
-  removeCard,
-  editProfile,
-  editUserAvatar,
-  getAllInfo,
-  changeLikeStatus,
-};
